+++ src/main/java/org/group2/webapp/service/ClaimService.java
@@ -41,18 +41,22 @@ public class ClaimService {
     @Transactional(readOnly = true)
     public Claim findOne(Long id) {
         log.debug("Request to get Claim : {}", id);
+        if (id == null)
+            return null;
         Claim claim = claimRepository.findOne(id);
         return claim;
     }
 
     public void delete(Long id) {
         log.debug("Request to delete Claim : {}", id);
+        if (id == null || claimRepository.findOne(id) == null)
+            return;
         claimRepository.delete(id);
     }
 
     public List<Claim> findClaimsPerFaculty(Long id) {
         log.debug("Request to get all Claims per Faculty : {}", id);
-       List<Claim> result = claimRepository.findAllByFacultyId(id);
+        List<Claim> result = claimRepository.findAllByFacultyId(id);
         return result;
     }
 }
